Solved data packing lab final final

+++ labs/memory_bound/data_packing/solution.h
@@ -12,7 +12,7 @@ struct S {
   float d;
   long long l : 16;
   int i : 8;
-  short s : 7;
+  unsigned s : 7;
   bool b : 1;
 
   bool operator<(const S &s) const { return this->i < s.i; }
